[neptune/#169] Block editing Verified Field Visits & allow deleting Assessments/Maintenance records Bug fix: hitting "Wrap up" at any time should redirect you to the detail page so you don't get an error
