+++ Source/Neptune.Web/Security/FieldVisitVerifyFeature.cs
@@ -1,0 +1,35 @@
+﻿using System.Collections.Generic;
+using Neptune.Web.Models;
+
+namespace Neptune.Web.Security
+{
+    [SecurityFeatureDescription("Allows verifying and unverifying Field Visits for a BMP if you are assigned to edit that BMP's jurisdiction")]
+    public class FieldVisitVerifyFeature : NeptuneFeatureWithContext, INeptuneBaseFeatureWithContext<FieldVisit>
+    {
+        private readonly NeptuneFeatureWithContextImpl<FieldVisit> _lakeTahoeInfoFeatureWithContextImpl;
+
+        public FieldVisitVerifyFeature()
+            : base(new List<Role> { Role.SitkaAdmin, Role.Admin, Role.JurisdictionManager })
+        {
+            _lakeTahoeInfoFeatureWithContextImpl = new NeptuneFeatureWithContextImpl<FieldVisit>(this);
+            ActionFilter = _lakeTahoeInfoFeatureWithContextImpl;
+        }
+
+        public void DemandPermission(Person person, FieldVisit contextModelObject)
+        {
+            _lakeTahoeInfoFeatureWithContextImpl.DemandPermission(person, contextModelObject);
+        }
+
+        public PermissionCheckResult HasPermission(Person person, FieldVisit contextModelObject)
+        {
+            var isAssignedToStormwaterJurisdiction = person.IsAssignedToStormwaterJurisdiction(contextModelObject.TreatmentBMP.StormwaterJurisdiction);
+
+            if (!isAssignedToStormwaterJurisdiction)
+            {
+                return new PermissionCheckResult($"You aren't assigned to edit Field Visit data for Jurisdiction {contextModelObject.TreatmentBMP.StormwaterJurisdiction.GetOrganizationDisplayName()}");
+            }
+
+            return new PermissionCheckResult();
+        }
+    }
+}
